Change restart policy to on-failure
The "always" restart policy "always restarts the container until its
removal." This means that that one could explicitly stop a container and
then restart one's computer, and the container would still be running
after the reboot.

Given that the Compose file in this repository appears to be used for
development and QA purposes, this is a bit painful. I'd rather not have
dev or qa tooling making my computer do things even after reboot.