fix(api): avoid panic when asserting x-rh-identity
The function identity.Get performs an unsafe type assertion on an
interface. If the underlying value of the interface does not assert to
the identity.XRHID type, the assertion fails and panics the program. To
avoid this panic, replace all calls to identity.Get with safe type
assertions on the interface directly. This allows us to gracefully
handle the case where the type assertion fails.