feat: add POST /manage HTTP API method
Extend the HTTP API with a "/manage" path that accepts an HTTP POST
method. The body of the request must be either true or false. Setting
the value to true will enable automatic configuration of state for all
connected clients (this is the current, default behavior). Setting the
value to false will set a flag on the database row that will be checked
when a host is created or updated. If the automatic management has been
disabled, then the host event is skipped and no configuration playbook
is sent.