test(inventory-consumer): remove handler_test.go
The handler.onMessage method is effectively the main routine of the
inventory-consumer process or goroutine. Unit-testing it adds
significant complexity to mock *most* of the external services that it
depends on, resulting in a mostly-controlled unit test with little
value. For the main process routines, it's fine to test them with
integratino tests alone. Focus unit-testing on the pieces that make up
the rest of the application.