feat(log): use level and formatting variables
Define two environment variables: CM_LOG_LEVEL and CM_LOG_FORMAT.

CM_LOG_LEVEL will be parsed by the logging library and set as the output
level. It defaults to "info".

CM_LOG_FORMAT can be used to set the output format. It defaults to
"json", but can be set to "text" to format output suitable for a
terminal emulator.