ci: use EXTRA_DEPLOY_ARGS to deploy mosquitto
In order for the entire live environment to run, we need to modify the
applications deployed. In addition to 'cloud-connector-api', we need to pull
in 'cloud-connector', which in turn brings in the 'mosquitto' component.
Additionally, override some of the default parameters that are set in the
ephemeral target with actual values.