refactor(db): change API to use package functions
Rather than passing around a handle, the db package exports functions
that all operate on a single, unexported handle. Importing packages can
then call the global functions directly, rather than needing to carry a
pointer to a db handle. The application doesn't need to ensure that it
initializes and passes around a handle.