feat(log): enable cloudwatch logging
Enable CloudWatch logging by using a MultiWriter as the log output
destination. The MultiWriter is configured during application startup
depending on the configuration values. If the application is running
with Clowder enabled, it attempts to create a cloudwatch.BatchWriter and
use that along with either the standard JSON writer or a ConsoleWriter.

cloudwatch.BatchWriter is a wrapper around the cloudwatchlogs API client
that implements io.Writer. BatchWriter is initialized with a batch
frequency, which is used to determine how often log events are sent to
the AWS CloudWatch Logs API.