refactor: use ffcli subcommands
Remove the bespoke "start" module func logic, instead using discrete
ffcli.Command variables as subcommands. The existing -module flag
behavior is maintained should the subcommands need to be run as separate
goroutines within a single process, but the ClowdApp definition
maintains the usage of separate pods/processes for each of the 3
subcommands in the application (http-api, inventory-consumer,
dispatcher-consumer).

The bespoke Kafka functions move from a top-level, public package to an
internal "util" package. The "domain/message" package is removed, with
its types defined near the handler that need them. A new "Host" type is
defined in the "internal" package that can be used (and will be used
more) to generically identify a host record, regardless of which
external service creates it.