fix: use per_reporter_staleness to detect reporter
The "reporter" field only contains the most recent reporter for a host.
In order to send a state playbook to a host, config-manager needs to
know if it was ever reported by "cloud-connector", rather than if it was
most recently reported. The "per_reporter_staleness" field contains a
map of all reporters for the host, so using the presence or absence of a
key in that map correctly identifies whether a host has ever been
reported by cloud-connector.