perf: Conditionally register insights-client
Only register insights-client if it's not already registered. This saves
CPU cycles and bandwidth on both client-side and server-side.

A properly registered host starts and enables `insights-client.timer`,
meaning that additional runs of `insights-client` are only necessary if
one expects new information from a client for a given run of
`insights-client`. This is the case after (say) a remediation execution,
but not after (re)starting rhcd.service or changing a server-side config
manager option.

Fix: https://issues.redhat.com/browse/ESSNTL-3626