fix: Ignore `insights-client --status` exit code
The exit code of `insights-client --status` varies:

* If `legacy_upload=False`, it will be 0 regardless of the registration
  status, and 1 if it encounters a problem while running.
* If `legacy_upload=True`, it will be 0 if registered, and 1 otherwise.

This means that relying on the exit code of `insights-client --status`
to decide whether to (re)register is chancy.

One solution is to unconditionally run `insights-client --register`,
which is inefficient, but a known-working solution. However, when
RHIN-500 is complete, re-registration will produce an error.

Another solution is to judge the meaning of `insights-client --status`
based not on its exit code, but on its stdout. Do so.