fix(inventory-consumer): don't ignore hosts
Relying on host reporter is unreliable; a host creation event might only
have puptoo as a reporter, even though cloud-connector appears as a
reporter shortly afterwards, in a host update event. Since there is only
ever a single host creation event, we must handle it, relying instead on
the checks in SetupHost that make a request to cloud-connector for the
current dispatchers map. Only if 'package-manager' is present and
'rhc-worker-playbook' is absent does the SetupHost operation proceed.