feat: add 'configuring' table
Add a 'configuring' table to the database. This table is inserted into
before the ApplyHost is initiated. If another host event is read off the
Kafka topic, the table first checked for the host ID; if a row exists
for the host ID and current profile ID, the ApplyHost routine is
skipped. When the dispatcherconsumer message reader reads a successful
message from playbook-dispatcher, it deletes the host ID/profile ID row
from the configuring table.